Add support for stdlib discriminator pattern
The Crystal stdlib serialization libraries (JSON and YAML) support the
concept of "discriminators" that dynamically choose the type of object
to deserialize to based on the value of a discriminator field. For
example:

    abstract struct Message
      include MessagePack::Serializable

      use_msgpack_discriminator "type", {
        created: Created,
        updated: Updated,
      }
    end

When the deserialization target is `Message`, if the `type` field is the
string "created", MessagePack will deserialize to a `Created` instance.